requirements_simple.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Data processing
PyYAML>=5.4.0
requests>=2.25.0
akshare>=1.10.0

# Machine learning
xgboost>=1.4.0
lightgbm>=3.2.0

# Development tools
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0

ruamel.yaml>=0.17.0 